opcodes: added a list of all the opcodes
To do this, we edited the python script that read the json. We also replaced the json (it was incorrect with the number of cycles that some commands took). This also added the dependency phf (perfect hash functions) to create static hash maps in compile time (regular hash maps are dynamically allocated, and can't be const in the current rust version)